Reorganise the tests
