+++ tests/triggers/TRPR0020.test.ts
@@ -1,10 +1,10 @@
 jest.setTimeout(30000)
 
-import { Guilt } from "../src/types/Guilt"
-import { TriggerCode } from "../src/types/TriggerCode"
-import generateMessage from "./helpers/generateMessage"
-import PostgresHelper from "./helpers/PostgresHelper"
-import processMessage from "./helpers/processMessage"
+import { Guilt } from "../../src/types/Guilt"
+import { TriggerCode } from "../../src/types/TriggerCode"
+import generateMessage from "../helpers/generateMessage"
+import PostgresHelper from "../helpers/PostgresHelper"
+import processMessage from "../helpers/processMessage"
 
 const code = TriggerCode.TRPR0020
 const resultCode = 3501
